// src/app/modules/common-regobs-api/model/attachment-edit-model.ts
/**
 * RegObs API, build: Local build - commit: ec80925f
 *
 * Contact: regobs@nve.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AttachmentEditModel { 
    /**
     * Navn på fotograf.
     */
    Photographer?: string | null;
    /**
     * Rettigheter til bilde.
     */
    Copyright?: string | null;
    /**
     * Hvilken himmelretning peker bilde. Gis i grader slik gitt på kompass. 0 er nord og 90 er øst osv.
     */
    Aspect?: number | null;
    /**
     * Sett naturfare. Tabellen brukes av alle naturfarer (snø, jord, vann, is). The GeoHazardKD unique identifier
     */
    GeoHazardTID?: number | null;
    /**
     * Hva er bildet av. Dette feltet relaterer bildet til en observasjonstype. Feks værobservasjon, faretegn, osv. The RegistrationKD unique identifier
     */
    RegistrationTID?: number | null;
    /**
     * Kommentarfelt for bildet. F.eks for å beskrive det.
     */
    Comment?: string | null;
    AttachmentMimeType?: string | null;
    /**
     * Om bildet skal vises først i registreringen, eller ikke
     */
    IsMainAttachment?: boolean | null;
    AttachmentId?: number | null;
    /**
     * Last opp attachment på forhånd og sett AttachmentUploadId
     */
    AttachmentUploadId?: string | null;
}